code for powerups
Time Warp:
Temporarily slows down time for the player who activates it, giving them more time to react to incoming balls. This can be a defensive move or used to set up precise shots.

Reverse Control:
Reverses the opponent's controls for a brief period, causing their paddle to move in the opposite direction. This power-up can be used strategically to create openings for scoring.

Time Bomb:
Attaches a time bomb to the ball, causing it to explode after a certain duration or upon impact with a paddle. The explosion can temporarily disable the opponent's paddle or create a shockwave affecting the ball's path