powerups and bugfix
CurveBall: Ball moves in a curving trajectory instead of straight.

Magnet Paddle: Attracts the ball towards the paddle for a brief period. I noticed this was pretty similar to Brice's "Homing", so instead of deleting it for now we can just scrap it when merging.

Time Warp: Slows down time for everything except the player's paddle.

ZigZagBall: Ball moves in a zigzag trajectory instead of straight.

BugFixes
- Fixed a bug where splitball clone will count for 50+ points.